scripts/requirements.txt
google-genai
pinecone-client==5.0.1
python-frontmatter==1.1.0
pypandoc==1.13
pymupdf4llm
mammoth
python-pptx